Snapshot-1
Changes up to rebuild...
Need to check the way sift_down_range works for binary heaps.
It's not clear this can be converted to an n-ary heap.
This affects:
1. rebuild
2. append
3. into_sorted_vec

If it can't be converted there are alternative implementations for
rebuild and append - but into_sorted_vec looks problematic - if it
really returns a sorted vec. In the worst case(?) we can use
sort_unstable_by_key on the data. This doesn't take advantage of the
known relationships in the heap, but given the main motivation for gheap
is to provide a VM friendly priority queue this use case doesn't appear
that significant.