Change generic parameters to enable the compiler to (continue to) type inference
to enable the compiler to (continue to) type inference from_vec_cmp.

But leaves issue of from_vec_indexer and from_vec_cmp_indexer (which
need tests).

Also: are the semantic of into_sorted_vec (and friends) questionable
give Min heap. And if so, what to do?

Plus benchmarks and stress testing needed....